fix: Prevent integer overflow in S3VolumeManager.getVolumeSizeLimit()
Fixes #209

Fixed integer overflow bug in getVolumeSizeLimit() by casting to long
before multiplication. Without this fix, volumes larger than 2GB would
overflow during the MB-to-bytes conversion.

Changes:
- Cast mount.getMaxSizeMB() to long before multiplication
- Use long literals (1024L) for multiplication operands
- Ensures calculation is done in long arithmetic from the start

Before this fix:
- 3GB volume: returned -1,149,239,296 (negative due to overflow)
- 5GB volume: returned 536,870,912 (wrapped around)

After this fix:
- 3GB volume: returns 3,145,728,000 (correct)
- 5GB volume: returns 5,242,880,000 (correct)

+++ jplatform-storage-s3/src/main/java/org/flossware/jplatform/storage/s3/S3VolumeManager.java
@@ -201,7 +201,7 @@ public class S3VolumeManager implements VolumeManager, AutoCloseable {
         if (mount == null) {
             throw new IllegalArgumentException("Volume not defined: " + volumeName);
         }
-        return mount.getMaxSizeMB() * 1024 * 1024;
+        return (long) mount.getMaxSizeMB() * 1024L * 1024L;
     }
 
     @Override
